backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
pandas==2.1.4
numpy==1.25.2
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
aiofiles==23.2.1
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
